pyproject.toml
[tool.poetry]
name = "rag_refapp"
version = "0.1.0"
description = ""
authors = ["Subodh Khanal <skhanal5@terpmail.umd.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
pre-commit = "^4.0.1"
httpx = "^0.27.2"
pytest = "^8.3.3"
opensearch-py = "^2.7.1"
haystack-ai = "^2.6.1"
opensearch-haystack = "^1.0.0"
sentence-transformers = "^3.2.0"
markdown-it-py = "^3.0.0"
mdit-plain = "^1.0.1"
pydantic-settings = "^2.6.0"
ruff = "^0.7.1"
mypy = "^1.13.0"

[tool.mypy]
plugins = ["pydantic.mypy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
